+++ packages/nextjs-router/src/legacy-app/index.ts
@@ -1,8 +1,0 @@
-/// <reference types="next" />
-/// <reference types="next/app" />
-
-export { routerProvider as default } from "./router-provider";
-export { NextRouteComponent } from "./route-component";
-
-export { RefineLink as Link } from "./../legacy-common/refine-link";
-export { handleRefineParams } from "./../legacy-common/handle-refine-params";