+++ packages/mui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @refinedev/mui
 
+## 5.14.4
+
+### Patch Changes
+
+-   [#5573](https://github.com/refinedev/refine/pull/5573) [`546df06482`](https://github.com/refinedev/refine/commit/546df06482807e59a7f2a735361a8e9169bb2563) Thanks [@alicanerdurmaz](https://github.com/alicanerdurmaz)! - chore: add "use client" directive to exported files to work with nextjs app router
+
+-   Updated dependencies [[`546df06482`](https://github.com/refinedev/refine/commit/546df06482807e59a7f2a735361a8e9169bb2563)]:
+    -   @refinedev/react-hook-form@4.8.14
+
 ## 5.14.3
 
 ### Patch Changes